Eliminated unnecessary code & refactored the code following a service for the GPT Facade.
