+++ assistant_id.json
@@ -1,5 +1,3 @@
 {
-    "assistant_id": "asst_lgASreK3QNvsDfQPckBXUNkB",
-    "json_assistant_id": "asst_qCN07O3f5khiUjOkr1vsaCFm",
     "agorai": "asst_OpBPSIckpk0ezCSz3nDfWL4Y"
 }
